requirements.txt
httpx~=0.28.1
pytest~=8.3.4
allure-python-commons~=2.13.5
python-dotenv~=1.0.1
config~=0.5.1
assertpy~=1.1
pip~=24.3.1
attrs~=25.1.0
wheel~=0.45.1
setuptools~=75.6.0
packaging~=24.2
typing_extensions~=4.12.2
python-dateutil~=2.9.0.post0